+++ app/src/main/java/com/kuit/afternote/data/repositoryimpl/MemorialThumbnailUploadRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.data.repositoryimpl
 
-import com.kuit.afternote.data.requireData
+import com.kuit.afternote.data.dto.response.requireData
 import com.kuit.afternote.domain.repository.MemorialThumbnailUploadRepository
 import com.kuit.afternote.feature.user.data.api.ImageApiService
 import com.kuit.afternote.feature.user.data.dto.PresignedUrlRequestDto
