+++ app/src/main/java/com/kuit/afternote/data/repositoryimpl/MemorialVideoUploadRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.kuit.afternote.data.repositoryimpl
 import android.content.Context
 import android.net.Uri
 import androidx.core.net.toUri
-import com.kuit.afternote.data.requireData
+import com.kuit.afternote.data.dto.response.requireData
 import com.kuit.afternote.domain.repository.MemorialVideoUploadRepository
 import com.kuit.afternote.feature.user.data.api.ImageApiService
 import com.kuit.afternote.feature.user.data.dto.PresignedUrlRequestDto
