+++ app/src/main/java/com/kuit/afternote/data/repositoryimpl/PhotoUploadRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.kuit.afternote.data.repositoryimpl
 
 import android.content.Context
 import android.net.Uri
-import com.kuit.afternote.data.requireData
+import com.kuit.afternote.data.dto.response.requireData
 import com.kuit.afternote.domain.repository.PhotoUploadRepository
 import com.kuit.afternote.feature.user.data.api.ImageApiService
 import com.kuit.afternote.feature.user.data.dto.PresignedUrlRequestDto
