+++ app/src/main/java/com/kuit/afternote/feature/admin/data/api/AdminApiService.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.admin.data.api
 
-import com.kuit.afternote.data.BaseResponse
+import com.kuit.afternote.data.dto.response.BaseResponse
 import com.kuit.afternote.feature.admin.data.dto.AdminVerificationResponseDto
 import com.kuit.afternote.feature.admin.data.dto.ApproveVerificationRequestDto
 import com.kuit.afternote.feature.admin.data.dto.RejectVerificationRequestDto
