+++ app/src/main/java/com/kuit/afternote/feature/admin/data/repository/impl/AdminRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.admin.data.repository.impl
 
-import com.kuit.afternote.data.requireData
+import com.kuit.afternote.data.dto.response.requireData
 import com.kuit.afternote.feature.admin.data.api.AdminApiService
 import com.kuit.afternote.feature.admin.data.dto.AdminVerificationResponseDto
 import com.kuit.afternote.feature.admin.data.dto.ApproveVerificationRequestDto
