+++ app/src/main/java/com/kuit/afternote/feature/receiver/data/api/ReceivedApiService.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.receiver.data.api
 
-import com.kuit.afternote.data.BaseResponse
+import com.kuit.afternote.data.dto.response.BaseResponse
 import com.kuit.afternote.feature.receiver.data.dto.CreateMindRecordReceiverRequestDto
 import com.kuit.afternote.feature.receiver.data.dto.CreateTimeLetterReceiverRequestDto
 import retrofit2.http.Body
