+++ app/src/main/java/com/kuit/afternote/feature/receiver/data/repository/impl/ReceivedRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.receiver.data.repository.impl
 
-import com.kuit.afternote.data.requireData
+import com.kuit.afternote.data.dto.response.requireData
 import com.kuit.afternote.feature.receiver.data.api.ReceivedApiService
 import com.kuit.afternote.feature.receiver.data.dto.CreateMindRecordReceiverRequestDto
 import com.kuit.afternote.feature.receiver.data.dto.CreateTimeLetterReceiverRequestDto
