+++ app/src/main/java/com/kuit/afternote/feature/receiverauth/data/api/ReceiverAuthApiService.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.receiverauth.data.api
 
-import com.kuit.afternote.data.BaseResponse
+import com.kuit.afternote.data.dto.response.BaseResponse
 import com.kuit.afternote.feature.receiverauth.data.dto.DeliveryVerificationRequestDto
 import com.kuit.afternote.feature.receiverauth.data.dto.DeliveryVerificationResponseDto
 import com.kuit.afternote.feature.receiverauth.data.dto.ReceivedAfternoteDetailAuthResponseDto
