+++ app/src/main/java/com/kuit/afternote/feature/timeletter/data/api/TimeLetterApiService.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.timeletter.data.api
 
-import com.kuit.afternote.data.BaseResponse
+import com.kuit.afternote.data.dto.response.BaseResponse
 import com.kuit.afternote.feature.timeletter.data.dto.TimeLetterCreateRequest
 import com.kuit.afternote.feature.timeletter.data.dto.TimeLetterDeleteRequest
 import com.kuit.afternote.feature.timeletter.data.dto.TimeLetterListResponse
