+++ app/src/main/java/com/kuit/afternote/feature/user/data/api/ImageApiService.kt
@@ -1,6 +1,6 @@
 package com.kuit.afternote.feature.user.data.api
 
-import com.kuit.afternote.data.BaseResponse
+import com.kuit.afternote.data.dto.response.BaseResponse
 import com.kuit.afternote.feature.user.data.dto.PresignedUrlRequestDto
 import com.kuit.afternote.feature.user.data.dto.PresignedUrlResponseDto
 import retrofit2.http.Body
