+++ app/src/main/java/com/kuit/afternote/feature/user/data/repository/UserRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.kuit.afternote.feature.user.data.repository
 
 import android.util.Log
-import com.kuit.afternote.data.requireData
-import com.kuit.afternote.data.requireStatus
+import com.kuit.afternote.data.dto.response.requireData
+import com.kuit.afternote.data.dto.response.requireStatus
 import com.kuit.afternote.data.service.ApiException
 import com.kuit.afternote.feature.user.data.api.UserApiService
 import com.kuit.afternote.feature.user.data.dto.RegisterReceiverRequestDto
